wip: 가게 상세화면 스크롤뷰로 변경중. StoreOverview, StoreInfoVIew 구현 완료

+++ 3dollar-in-my-pocket/domain/home/subviews/cells/StoreCell.swift
@@ -147,7 +147,7 @@ class StoreCell: BaseCollectionViewCell {
     }
   }
   
-  private func setRating(rating: Float) {
+  private func setRating(rating: Double) {
     if floor(rating) == rating {
       self.rankLabel.text = "\(Int(rating))점"
     } else {
